+++ examples/device_info.rs
@@ -3,14 +3,7 @@ use std::collections::{BTreeSet, HashMap};
 use shiguredo_video_device::{PixelFormat, VideoDeviceList};
 
 fn main() {
-    #[cfg(all(target_os = "linux", feature = "v4l2"))]
-    let device_list = VideoDeviceList::enumerate_v4l2();
-    #[cfg(all(target_os = "linux", feature = "pipewire", not(feature = "v4l2")))]
-    let device_list = VideoDeviceList::enumerate_pipewire();
-    #[cfg(target_os = "macos")]
-    let device_list = VideoDeviceList::enumerate_avf();
-    #[cfg(target_os = "windows")]
-    let device_list = VideoDeviceList::enumerate_mf();
+    let device_list = VideoDeviceList::enumerate();
 
     let device_list = match device_list {
         Ok(list) => list,